Keep runtime loader stream-status logic easier to extend
Extract the tool-input status transition logic into a dedicated helper
module so runtime-loader.ts stays focused on provider request/response
translation instead of stream heartbeat bookkeeping.

Constraint: Preserve runtime loader streaming semantics and exported API surface
Constraint: Repo pre-commit is currently blocked by pre-existing CLI boundary violations unrelated to this change
Rejected: Splitting broader request parsing helpers in the same change | would widen scope beyond one queue item
Confidence: high
Scope-risk: narrow
Directive: Keep tool-input-status.ts limited to stream status transitions; leave provider-specific parsing in runtime-loader.ts
Tested: deno test --no-check --allow-all --unstable-worker-options src/provider/runtime-loader.test.ts; deno fmt --check src/provider/runtime-loader.ts src/provider/runtime-loader/tool-input-status.ts; deno lint src/provider/runtime-loader.ts src/provider/runtime-loader/tool-input-status.ts; git diff --check
Not-tested: Pre-commit verify:quick (fails on existing cli boundary violations in cli/commands/extension/*.ts)
